+++ Assets/_scripts/HighScore/Score.cs
@@ -1,7 +1,7 @@
 ﻿public class Score
 {
     public string UserName { get; set; }
-    public float Value { get; set; }
+    public int Value { get; set; }
 
     public Score()
     {
